Make entire oscillator shader highp

+++ src/shaders/glsl/oscillator.glsl
@@ -1,4 +1,4 @@
-precision mediump float;
+precision highp float;
 
 #pragma glslify: sampleTex = require('./sampleTex');
 
@@ -77,12 +77,14 @@ void main() {
 		sampleTex(shape, textureSamplePoint, shapeAmount);
 
 	float duty = clamp(sampledShape, 0.5, 1.);
-	// should mod pixelIndex by 1, but guaranteed to be within 0-1
 	float p1 = duty;
 	float p2 = 1. - p1;
-	highp float m1 = 0.5 / p1;
-	highp float m2 = 0.5 / p2;
+	float m1 = 0.5 / p1;
+	float m2 = 0.5 / p2;
+
+	// should mod pixelIndex by 1, but guaranteed to be within 0-1
 	float x = mod((frequency * pixelIndex) + phaseOffsetFromTexture - 0.25, 1.);
+
 	float inFirstDuty = 1. - step(p1, x);
 
 	float t =
